+++ src/monitor-control-service/azext_amcs/vendored_sdks/amcs/aio/operations/_data_collection_rule_associations_operations.py
@@ -26,7 +26,7 @@ class DataCollectionRuleAssociationsOperations:
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure.mgmt.amcs.models
+    :type models: ~$(python-base-namespace).v2021_09_01_preview.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -54,7 +54,7 @@ class DataCollectionRuleAssociationsOperations:
         :type resource_uri: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either DataCollectionRuleAssociationProxyOnlyResourceListResult or the result of cls(response)
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.amcs.models.DataCollectionRuleAssociationProxyOnlyResourceListResult]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResourceListResult]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["models.DataCollectionRuleAssociationProxyOnlyResourceListResult"]
@@ -62,7 +62,7 @@ class DataCollectionRuleAssociationsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-04-01"
+        api_version = "2021-09-01-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -102,7 +102,7 @@ class DataCollectionRuleAssociationsOperations:
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
-                error = self._deserialize(models.ErrorResponse, response)
+                error = self._deserialize(models.ErrorResponseCommonV2, response)
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
@@ -130,7 +130,7 @@ class DataCollectionRuleAssociationsOperations:
         :type data_collection_rule_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either DataCollectionRuleAssociationProxyOnlyResourceListResult or the result of cls(response)
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.amcs.models.DataCollectionRuleAssociationProxyOnlyResourceListResult]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResourceListResult]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["models.DataCollectionRuleAssociationProxyOnlyResourceListResult"]
@@ -138,7 +138,7 @@ class DataCollectionRuleAssociationsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-04-01"
+        api_version = "2021-09-01-preview"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -151,7 +151,7 @@ class DataCollectionRuleAssociationsOperations:
                 url = self.list_by_rule.metadata['url']  # type: ignore
                 path_format_arguments = {
                     'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
-                    'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1, pattern=r'^[-\w\._\(\)]+$'),
+                    'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
                     'dataCollectionRuleName': self._serialize.url("data_collection_rule_name", data_collection_rule_name, 'str'),
                 }
                 url = self._client.format_url(url, **path_format_arguments)
@@ -180,7 +180,7 @@ class DataCollectionRuleAssociationsOperations:
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
-                error = self._deserialize(models.ErrorResponse, response)
+                error = self._deserialize(models.ErrorResponseCommonV2, response)
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
@@ -190,6 +190,84 @@ class DataCollectionRuleAssociationsOperations:
             get_next, extract_data
         )
     list_by_rule.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Insights/dataCollectionRules/{dataCollectionRuleName}/associations'}  # type: ignore
+
+    def list_by_data_collection_endpoint(
+        self,
+        resource_group_name: str,
+        data_collection_endpoint_name: str,
+        **kwargs
+    ) -> AsyncIterable["models.DataCollectionRuleAssociationProxyOnlyResourceListResult"]:
+        """Lists associations for the specified data collection endpoint.
+
+        Lists associations for the specified data collection endpoint.
+
+        :param resource_group_name: The name of the resource group. The name is case insensitive.
+        :type resource_group_name: str
+        :param data_collection_endpoint_name: The name of the data collection endpoint. The name is
+         case insensitive.
+        :type data_collection_endpoint_name: str
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :return: An iterator like instance of either DataCollectionRuleAssociationProxyOnlyResourceListResult or the result of cls(response)
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResourceListResult]
+        :raises: ~azure.core.exceptions.HttpResponseError
+        """
+        cls = kwargs.pop('cls', None)  # type: ClsType["models.DataCollectionRuleAssociationProxyOnlyResourceListResult"]
+        error_map = {
+            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
+        }
+        error_map.update(kwargs.pop('error_map', {}))
+        api_version = "2021-09-01-preview"
+        accept = "application/json"
+
+        def prepare_request(next_link=None):
+            # Construct headers
+            header_parameters = {}  # type: Dict[str, Any]
+            header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
+
+            if not next_link:
+                # Construct URL
+                url = self.list_by_data_collection_endpoint.metadata['url']  # type: ignore
+                path_format_arguments = {
+                    'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str', min_length=1),
+                    'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', max_length=90, min_length=1),
+                    'dataCollectionEndpointName': self._serialize.url("data_collection_endpoint_name", data_collection_endpoint_name, 'str'),
+                }
+                url = self._client.format_url(url, **path_format_arguments)
+                # Construct parameters
+                query_parameters = {}  # type: Dict[str, Any]
+                query_parameters['api-version'] = self._serialize.query("api_version", api_version, 'str')
+
+                request = self._client.get(url, query_parameters, header_parameters)
+            else:
+                url = next_link
+                query_parameters = {}  # type: Dict[str, Any]
+                request = self._client.get(url, query_parameters, header_parameters)
+            return request
+
+        async def extract_data(pipeline_response):
+            deserialized = self._deserialize('DataCollectionRuleAssociationProxyOnlyResourceListResult', pipeline_response)
+            list_of_elem = deserialized.value
+            if cls:
+                list_of_elem = cls(list_of_elem)
+            return deserialized.next_link or None, AsyncList(list_of_elem)
+
+        async def get_next(next_link=None):
+            request = prepare_request(next_link)
+
+            pipeline_response = await self._client._pipeline.run(request, stream=False, **kwargs)
+            response = pipeline_response.http_response
+
+            if response.status_code not in [200]:
+                error = self._deserialize(models.ErrorResponseCommonV2, response)
+                map_error(status_code=response.status_code, response=response, error_map=error_map)
+                raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
+
+            return pipeline_response
+
+        return AsyncItemPaged(
+            get_next, extract_data
+        )
+    list_by_data_collection_endpoint.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Insights/dataCollectionEndpoints/{dataCollectionEndpointName}/associations'}  # type: ignore
 
     async def get(
         self,
@@ -207,7 +285,7 @@ class DataCollectionRuleAssociationsOperations:
         :type association_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: DataCollectionRuleAssociationProxyOnlyResource, or the result of cls(response)
-        :rtype: ~azure.mgmt.amcs.models.DataCollectionRuleAssociationProxyOnlyResource
+        :rtype: ~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResource
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["models.DataCollectionRuleAssociationProxyOnlyResource"]
@@ -215,7 +293,7 @@ class DataCollectionRuleAssociationsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-04-01"
+        api_version = "2021-09-01-preview"
         accept = "application/json"
 
         # Construct URL
@@ -240,7 +318,7 @@ class DataCollectionRuleAssociationsOperations:
 
         if response.status_code not in [200]:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = self._deserialize(models.ErrorResponse, response)
+            error = self._deserialize(models.ErrorResponseCommonV2, response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         deserialized = self._deserialize('DataCollectionRuleAssociationProxyOnlyResource', pipeline_response)
@@ -267,10 +345,10 @@ class DataCollectionRuleAssociationsOperations:
         :param association_name: The name of the association. The name is case insensitive.
         :type association_name: str
         :param body: The payload.
-        :type body: ~azure.mgmt.amcs.models.DataCollectionRuleAssociationProxyOnlyResource
+        :type body: ~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResource
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: DataCollectionRuleAssociationProxyOnlyResource, or the result of cls(response)
-        :rtype: ~azure.mgmt.amcs.models.DataCollectionRuleAssociationProxyOnlyResource
+        :rtype: ~$(python-base-namespace).v2021_09_01_preview.models.DataCollectionRuleAssociationProxyOnlyResource
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["models.DataCollectionRuleAssociationProxyOnlyResource"]
@@ -278,7 +356,7 @@ class DataCollectionRuleAssociationsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-04-01"
+        api_version = "2021-09-01-preview"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -311,7 +389,7 @@ class DataCollectionRuleAssociationsOperations:
 
         if response.status_code not in [200, 201]:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = self._deserialize(models.ErrorResponse, response)
+            error = self._deserialize(models.ErrorResponseCommonV2, response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if response.status_code == 200:
@@ -350,7 +428,7 @@ class DataCollectionRuleAssociationsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-04-01"
+        api_version = "2021-09-01-preview"
         accept = "application/json"
 
         # Construct URL
@@ -375,7 +453,7 @@ class DataCollectionRuleAssociationsOperations:
 
         if response.status_code not in [200, 204]:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = self._deserialize(models.ErrorResponse, response)
+            error = self._deserialize(models.ErrorResponseCommonV2, response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if cls: