+++ src/aks-preview/HISTORY.md
@@ -3,9 +3,15 @@
 Release History
 ===============
 
-0.5.64
+0.5.65
 ++++++
 * Add `--ignore-pod-disruption-budget` flag for `az aks nodepool delete` for ignoring PodDisruptionBudget.
+
+0.5.64
+++++++
+
+* Add support for updating kubelet identity. Command is
+  * `az aks update --assign-kubelet-identity <kubelelt-identity-resource-id>`
 
 0.5.63
 ++++++